fix: conditional reconciliation block, remove time_zone, move/rename functions
